trans: always register an item's symbol, even if duplicated.
